Move GetabIndex to `utils` file so symbols do not clash
